+++ CHANGELOG.md
@@ -14,17 +14,27 @@
 - **11/14** - hosted fonts from adobe added to project
 - **11/14** - began set up of game interface by defining game container and drawing the first graphic inside index.js
 - **11/14** - added typescript to the project compile-time error checking
-- **11/14** - completed basic ping pong animation for app 
+- **11/14** - completed basic ping pong animation for app
+- **11/16** - fully implemented border change logic on collision
+- **11/16** - fully implemented promise when all 4 borders are collided with to end game
+- **11/16** - added mesh gradient background and /assets folder inside src to store
+- **11/16** - finalized css
 
 **CHANGED:**
 
 - **11/14** - used pixi.js cdn link to import into project
-- **11/14** - index.js to index.ts for typescript 
+- **11/14** - index.js to index.ts for typescript
+- **11/16** - added playAgainBtn to restart game after promise is fulfilled
 
 **FIXED:**
+
+- **N/A**
 
 **REMOVED:**
 
 - **11/14** - package files and node modules folder after changing pixi import
 
 **RESOURCES USED:**
+
+- **border collision guidance:** https://pixijs.com/8.x/playground?exampleId=advanced.collisionDetection
+- **using void in promises:** https://typescript.tv/new-features/the-void-operator-in-typescript-and-javascript/